initial draft of splitting current user class into its own class, testing bugs
