+++ tests/fixtures/current_user.json
@@ -1,0 +1,61 @@
+{
+	"get_by_id": {
+		"method": "GET",
+		"endpoint": "users/self",
+		"data": {
+			"id": 1,
+			"name": "John Doe"
+		},
+		"status_code": 200
+	},
+	"list_groups": {
+		"method": "GET",
+		"endpoint": "users/self/groups",
+		"data": [
+			{
+				"id": 1,
+				"name": "Group 1"
+			},
+			{
+				"id": 2,
+				"name": "Group 2"
+			}
+		],
+		"status_code": 200,
+		"headers": {
+			"Link": "<http://example.com/api/v1/users/self/groups?page=2&per_page=2>; rel=\"next\""
+		}
+	},
+	"list_groups2": {
+		"method": "GET",
+		"endpoint": "users/self/groups?page=2&per_page=2",
+		"data": [
+			{
+				"id": 3,
+				"name": "Group 3"
+			},
+			{
+				"id": 4,
+				"name": "Group 4"
+			}
+		],
+		"status_code": 200
+	},
+	"list_calendar_events_for_user": {
+		"method": "GET",
+		"endpoint": "users/1/calendar_events",
+		"data": [
+			{
+				"id": 123,
+				"context_code": "course_321",
+				"title": "Test Event 1"
+			},
+			{
+				"id": 456,
+				"context_code": "course_654",
+				"title": "Test Event 2"
+			}
+		],
+		"status_code": 200
+	}
+}
